app/requirements.txt
Flask
mysql-connector-python
prometheus_client
gunicorn